pip-requirements.txt
python-telegram-bot==12.3.0

# pip3 install -r pip-requirements.txt